Moving find X in request to helpers
